refactor(migrate): rename URL migration scripts
- Rename URL migration entrypoints to migrate-urls and backup-urls
- Rename analyzer module to lib/analyze-urls
- Make createBackup accept file prefix and data payload
- Update URL migration guide and examples to new names
