+++ scripts/migrate/lib/analyze-urls.js
